+++ R/tables.R
@@ -1,3 +1,65 @@
+#' Copy BigQuery table
+#' 
+#' Copy a source table to another destination
+#' 
+#' @param source_projectid source table's projectId
+#' @param source_datasetid source table's datasetId
+#' @param source_tableid source table's tableId
+#' @param destination_projectid destination table's projectId
+#' @param destination_datasetid destination table's datasetId
+#' @param destination_tableid destination table's tableId
+#' @param createDisposition Create table's behaviour
+#' @param writeDisposition Write to an existing table's behaviour
+#' 
+#' @return A job object
+#' 
+#' @export
+#' @import assertthat
+bqr_copy_table <- function(source_tableid,
+                           destination_tableid,
+                           source_projectid = bqr_get_global_project(),
+                           source_datasetid = bqr_get_global_dataset(),
+                           destination_projectid = bqr_get_global_project(),
+                           destination_datasetid = bqr_get_global_dataset(),
+                           createDisposition = c("CREATE_IF_NEEDED","CREATE_NEVER"),
+                           writeDisposition = c("WRITE_TRUNCATE", "WRITE_APPEND", "WRITE_EMPTY")){
+  
+  createDisposition <- match.arg(createDisposition)
+  writeDisposition <- match.arg(writeDisposition)
+  
+  assert_that(
+    is.string(source_projectid),
+    is.string(source_datasetid),
+    is.string(source_tableid),
+    is.string(destination_projectid),
+    is.string(destination_datasetid),
+    is.string(destination_tableid)
+  )
+  
+  config <- list(
+    configuration = list(
+      copy = list(
+        createDisposition = createDisposition,
+        sourceTable = list(
+          projectId = source_projectid,
+          datasetId = source_datasetid,
+          tableId = source_tableid
+        ),
+        destinationTable = list(
+          projectId = destination_projectid,
+          datasetId = destination_datasetid,
+          tableId = destination_tableid
+        ),
+        writeDisposition = writeDisposition
+      )
+    )
+  )
+  
+  call_job(source_projectid, config = config)
+}
+
+
+
 #' List BigQuery tables in a dataset
 #' 
 #' @param projectId The BigQuery project ID