+++ NEWS.md
@@ -1,9 +1,12 @@
 # bigQueryR 0.3.2.9000
 
-* support `nullMarker`, `maxBadRecords` in upload jobs
+* support `nullMarker`, `maxBadRecords`, `fieldDelimiter` in upload jobs
 * Support BigQuery type `DATE` for R class `Date` data.frame columns (BigQuery type `TIMESTAMP` still default for `POSIXct`columns) (#48)
 * Allow custom user schema for uploads of data.frames (#48)
 * Rename misnamed global functions from `bq_` prefix to `bqr_` prefix
+* Add `allowJaggedRows` and `allowQuotedNewlines` options to upload via `bqr_upload_data()`
+* `bqr_get_job` now accepts a job object as well as the jobId
+* Fix bug with `bqr_upload_data` where `autodetect=TRUE` didn't work with `gcs://` loads from Cloud Storage
 
 # bigQueryR 0.3.2